一部の変数をdoubleに変更 移動にデルタ時間が考慮されていなかった問題を修正

+++ src/medipro/object/camera/FollowingCameraModel.java
@@ -10,8 +10,8 @@ public class FollowingCameraModel extends CameraModel {
 
     public Optional<GameObjectModel> target = Optional.empty();
 
-    public int originX = 0;
-    public int originY = 0;
+    public double originX = 0;
+    public double originY = 0;
 
     public FollowingCameraModel(World world, GameObjectModel target) {
         super(world);
