+++ src/medipro/object/player/PlayerView.java
@@ -38,10 +38,10 @@ public class PlayerView extends GameObjectView {
             AffineTransform tx = AffineTransform.getScaleInstance(-1, 1);
             tx.translate(-image.getWidth(null), 0);
             AffineTransformOp op = new AffineTransformOp(tx, AffineTransformOp.TYPE_NEAREST_NEIGHBOR);
-            g.drawImage(op.filter((BufferedImage) image, null), playerModel.x - SPRITE_WIDTH / 2,
-                    playerModel.y - SPRITE_HEIGHT, null);
+            g.drawImage(op.filter((BufferedImage) image, null), (int) (playerModel.x - SPRITE_WIDTH / 2),
+                    (int) (playerModel.y - SPRITE_HEIGHT), null);
         } else {
-            g.drawImage(image, playerModel.x - SPRITE_WIDTH / 2, playerModel.y - SPRITE_HEIGHT, null);
+            g.drawImage(image, (int) (playerModel.x - SPRITE_WIDTH / 2), (int) (playerModel.y - SPRITE_HEIGHT), null);
         }
     }
 }
